+++ include/Vector.h
@@ -10,21 +10,21 @@ namespace penguin {
     class Vector : public Point {
         public:
 
-        Vector(double xv = 0, double yv = 0);
+        Vector(float xv = 0, float yv = 0);
 
         Vector(const Point&, const Point&);
 
         // Angle With Vector (0, 0)
-        double Angle() const;
+        float Angle() const;
 
         // Angle Between This Vector And Another
-        double Angle(const Vector&) const;
+        float Angle(const Vector&) const;
         
         // Cross Product
         Vector operator* (const Vector&);
         
         // Dot Product
-        double operator^ (const Vector&) const;
+        float operator^ (const Vector&) const;
                 
     };
