+++ src/Vector.cpp
@@ -6,15 +6,15 @@
 
 namespace penguin {
 
-    Vector::Vector(double xv, double yv) : Point(xv, yv) {}
+    Vector::Vector(float xv, float yv) : Point(xv, yv) {}
 
     Vector::Vector(const Point& A, const Point& B) : Point(B.x - A.x, B.y - A.y) {}
 
-    double Vector::Angle() const {
+    float Vector::Angle() const {
         return atan(y / x) + (x < 0 ? PI : 0);
     }
 
-    double Vector::Angle(const Vector& V) const {
+    float Vector::Angle(const Vector& V) const {
         auto lu = this->Length();
         auto lv = V.Length();
         auto ds = lu * lv;
@@ -27,7 +27,7 @@ namespace penguin {
         return Vector { this->x * V.y, this->y * V.x };
     }
 
-    double Vector::operator^ (const Vector& V) const {
+    float Vector::operator^ (const Vector& V) const {
         return this->x * V.x + this->y * V.y;
     }
